Generate server (#5)
* Add OpenAPI Generator code to generate service

This is based on the example at:
https://www.baeldung.com/java-openapi-generator-server

* Exclude unwanted files from source control

* Exclude README from code generation

We want to keep our own README rather than overwriting it with the
generic generated one.

* Add generated files to source control

The generated Java server's endpoints all return with a default 501 (not
implemented) response, currently.  In a future commit we will implement
the endpoints needed to verify the service's contract with its consumer.