Add CI "can I deploy" check before deploying (#14)
This commit adds a job to the CI pipeline which checks that it is safe
to deploy to our integrated environments (pre-production and production)
before deploying.  The check is done by running the contract tests on
the specific commit on the contract repo that is associated (tagged)
with our consumer commit that is currently in the integrated environment
that we want to deploy to.  The contract tests run against the provider
code that we want to deploy, [using Prism as a validation proxy][1]. We
do not use the real integrated provider environment for these tests.
Instead we use our `ephemeral-petstore-service` Heroku environment
(which isn't really ephemeral, but serves the purpose by highlighting
that it would be ephemeral in a more "productionised" example).  We
could experiment in a later commit with using Heroku's [Review App
API][2] to create a genuinely ephemeral environment, but that's for a
later date (post-MVP).

[1]: https://devcenter.heroku.com/articles/github-integration-review-apps#review-apps-api

[2]: https://meta.stoplight.io/docs/prism/docs/guides/03-validation-proxy.md