Fix pipeline deployment for `can-i-deploy` check (#15)
The pipeline deployment was broken because we were trying to deploy to
the same environment that the consumer pipeline already uses. This led
to Heroku rejecting the deployment because the commit being pushed
was behind the remote branch.  The quick and dirty fix in this commit
is to use another environment that we created in Heroku, called
`ephemeral-petstore-service2`.