Run Portman tests on every pull request
The tests are automatically generated on the fly by Portman, which is
very useful indeed.  These tests verify that our provider implementation
is in sync with our OpenAPI specification.  This is just a MVP of
integrating Portman tests into our CI/CD pipeline at this stage.  We may
well be able to add more Portman tests (e.g. unhappy paths as well as
happy paths), but this is a good start.