Add Portman OpenAPI CLI report (#23)
This commit adds a very useful report to the CLI output when running
Portman.  This [Newman OpenAPI reporter][1] breaks down the test
coverage into the individual paths in the OpenAPI spec, making it clear
which pahts are not fully covered by the tests.

NB the reporter does not yet have functionality to output this report
in a file format (e.g. json or junit), but that is coming soon according
to the repo's README.  When that functionality arrives we can add it in
a future PR.

[1]: https://github.com/NickHeap2/newman-reporter-openapi