Add content-type Portman tests
We now check the content-type in the Portman tests that we automatically
generate.  See [the Portman contract tests documentation][1] for more
info.  To make the tests pass we have removed the unused xml
content-type in our OpenAPI spec when we encountered a failing test. It
is safe to do so as we know (via our consumer contract tests) that our
consumer only requires json as the content-type, not xml.

[1]: https://github.com/apideck-libraries/portman/tree/main/examples/testsuite-contract-tests#portman---contracttests-properties