Add first Portman test that covers an unhappy path
This commit adds a [Portman variation test][1] to cover just one of our
unhappy paths, as an example of how to do it (we can add more unhappy
path coverage in a future commit).  In our example here we use Portman
to remove the `name` property from the pet request body when trying to
update a Pet's details in the petstore, and verify that the service
responds with a `BAD_REQUEST` `400` response.

[1]: https://github.com/apideck-libraries/portman/tree/main/examples/testsuite-variation-tests